+++ Entities/Event.cs
@@ -14,7 +14,7 @@ namespace EventbriteNET.Entities
         public DateTime Created;
         public DateTime Modified;
 
-        public Dictionary<int, Ticket> Tickets = new Dictionary<int, Ticket>();
+        public Dictionary<long, Ticket> Tickets = new Dictionary<long, Ticket>();
 
         private List<Attendee> attendees;
         public List<Attendee> Attendees
